+++ liana/multi/to_mudata.py
@@ -1,40 +1,43 @@
+from __future__ import annotations
+from typing import Callable
+
 import numpy as np
 import pandas as pd
-import scanpy as sc
 import warnings as warnings
 
 from anndata import AnnData
+from mudata import MuData
 from tqdm import tqdm
 
 from ._common import _process_scores
 from liana._logging import _check_if_installed
+from liana._constants._docs import d
+from liana.method._pipe_utils import _check_groupby
 
-def adata_to_views(adata,
-                   groupby,
-                   sample_key,
-                   obs_keys,
-                   view_separator=':',
-                   min_count=10,
-                   min_total_count=15,
-                   large_n=10, 
-                   min_prop=0.1,
-                   keep_stats = False,
-                   verbose=False,
+@d.dedent
+def adata_to_views(adata: AnnData,
+                   groupby: str,
+                   sample_key: str,
+                   obs_keys: list = None,
+                   view_sep: str = ':',
+                   min_count: int = 10,
+                   min_total_count: int = 15,
+                   large_n: int = 10, 
+                   min_prop: float = 0.1,
+                   keep_stats: bool = False,
+                   verbose: bool = False,
                    **kwargs):
     """
     Converts an AnnData object to a MuData object with views that represent an aggregate for each entity in `adata.obs[groupby]`.
     
     Parameters
     ----------
-    adata: :class:`~anndata.AnnData`
-        AnnData object
-    groupby:
-        Column name in `adata.obs` to group by
-    sample_key:
-        Column name in `adata.obs` to use as sample key
+    %(adata)s
+    %(groupby)s
+    %(sample_key)s
     obs_keys:
         Column names in `adata.obs` to merge with the MuData object
-    view_separator:
+    view_sep:
         Separator to use when assigning `adata.var_names` to views
     min_count:
         Minimum number of counts per gene per sample to be included in the pseudobulk.
@@ -46,8 +49,7 @@ def adata_to_views(adata,
         Minimum proportion of samples that must have a count for a gene to be included in the pseudobulk.
     keep_stats:
         If True, keep the pseudobulk statistics in `mdata.uns['psbulk_stats']`. Default is False.
-    verbose:
-        If True, show progress bar.
+    %(verbose)s
     **kwargs
         Keyword arguments used to aggregate the values per cell into views. See `dc.filter_by_expr` for more details.
     
@@ -58,11 +60,12 @@ def adata_to_views(adata,
     """
     
     # Check if MuData & decoupler are installed
-    mu = _check_if_installed(package_name="mudata")
     dc = _check_if_installed(package_name="decoupler")
     
     views = adata.obs[groupby].unique()
     views = tqdm(views, disable=not verbose)
+    
+    _check_groupby(adata=adata, groupby=groupby, verbose=verbose)
     
     padatas = {}
     if keep_stats: stats = []
@@ -70,7 +73,7 @@ def adata_to_views(adata,
         # filter AnnData to view
         temp = adata[adata.obs[groupby] == view].copy()
         # assign view to var_names
-        temp.var_names = view + view_separator + temp.var_names
+        temp.var_names = view + view_sep + temp.var_names
 
         padata = dc.get_pseudobulk(temp,
                                    sample_col=sample_key,
@@ -96,14 +99,14 @@ def adata_to_views(adata,
             # keep psbulk stats
             if keep_stats:
                 df = padata.obs.filter(items=['psbulk_n_cells', 'psbulk_counts'], axis=1)
-                df.columns = [view + view_separator + col for col in df.columns]
+                df.columns = [view + view_sep + col for col in df.columns]
                 stats.append(df)
 
             del padata.obs
             padatas[view] = padata
 
     # Convert to MuData
-    mdata = mu.MuData(padatas)
+    mdata = MuData(padatas)
     
     # process metadata
     _process_meta(adata=adata, mdata=mdata, sample_key=sample_key, obs_keys=obs_keys)
@@ -114,41 +117,36 @@ def adata_to_views(adata,
     
     return mdata
 
-
-def lrs_to_views(adata, 
-                 score_key=None, 
-                 inverse_fun= lambda x: 1 - x,
-                 obs_keys=None,
-                 lr_prop=0.5,
-                 lr_fill=np.nan,
-                 lrs_per_view=20,
-                 lrs_per_sample=10,
-                 samples_per_view=3,
-                 min_variance=0,
-                 lr_separator='^',
-                 cell_separator='&',
-                 var_separator=':',
-                 uns_key = 'liana_res',
-                 sample_key='sample',
-                 source_key='source',
-                 target_key='target', 
-                 ligand_key='ligand_complex',
-                 receptor_key='receptor_complex',
-                 verbose=False
+@d.dedent
+def lrs_to_views(adata: AnnData, 
+                 score_key: (str or None) = None, 
+                 inverse_fun: Callable = lambda x: 1 - x,
+                 obs_keys: (list or None) =None,
+                 lr_prop: float = 0.5,
+                 lr_fill: np.nan = np.nan,
+                 lrs_per_view:int = 20,
+                 lrs_per_sample:int = 10,
+                 samples_per_view: int = 3,
+                 min_variance:int = 0,
+                 lr_sep:str='^',
+                 cell_sep:str='&',
+                 var_sep:str=':',
+                 uns_key:str = 'liana_res',
+                 sample_key:str='sample',
+                 source_key:str='source',
+                 target_key:str='target', 
+                 ligand_key:str='ligand_complex',
+                 receptor_key:str='receptor_complex',
+                 verbose:bool=False
                  ):
     """
     Converts a LIANA result to a MuData object with views that represent an aggregate for each entity in `adata.obs[groupby]`.
     
     Parameters
     ----------
-    adata
-        AnnData object with LIANA results in `adata.uns[uns_key]`
-    score_key
-        Column in `adata.uns[uns_key]` that contains the scores to be used for building the views.
-    inverse_fun
-        Function that is applied to the scores before building the views. Default is `lambda x: 1 - x` which is used to invert the scores
-        reflect probabilities (e.g. magnitude_rank), i.e. such for which lower values reflect higher relevance.
-        This is handled automatically for the scores in liana.
+    %(adata)s
+    %(score_key)s
+    %(inverse_fun)s
     obs_keys
         List of keys in `adata.obs` that should be included in the MuData object. Default is `None`. 
         These columns should correspond to the number of samples in `adata.obs[sample_key]`.
@@ -165,26 +163,18 @@ def lrs_to_views(adata,
     min_variance
         Reflects the minimum required variance across samples for each interaction in each view. Default is `0`.
         NaNs are ignored when computing the variance.
-    lr_separator
-        Separator to use for the interaction names in the views. Default is `^`.
-    cell_separator
+    %(lr_sep)s
+    cell_sep
         Separator to use for the cell names in the views. Default is `&`.
-    var_separator
+    var_sep
         Separator to use for the variable names in the views. Default is `:`.
-    uns_key
-        Key in `adata.uns` that contains the LIANA results. Default is `'liana_res'`.
-    sample_key
-        Key in `adata.uns[uns_key]` that contains the sample names. Default is `'sample'`.
-    source_key
-        Key in `adata.uns[uns_key]` that contains the source names. Default is `'source'`.
-    target_key
-        Key in `adata.uns[uns_key]` that contains the target names. Default is `'target'`.
-    ligand_key
-        Key in `adata.uns[uns_key]` that contains the ligand names. Default is `'ligand_complex'`.
-    receptor_key
-        Key in `adata.uns[uns_key]` that contains the receptor names. Default is `'receptor_complex'`.
-    verbose
-        If True, show progress bar.
+    %(uns_key)s
+    %(sample_key)s
+    %(source_key)s
+    %(target_key)s
+    %(ligand_key)s
+    %(receptor_key)s
+    %(verbose)s
     
     Returns
     -------
@@ -220,8 +210,8 @@ def lrs_to_views(adata,
         raise ValueError(f'`{missing_keys}` not found in `adata.uns[{uns_key}]`! Please check your input.')
     
     # concat columns (needed for MOFA)
-    liana_res['interaction'] = liana_res[ligand_key] + lr_separator + liana_res[receptor_key]
-    liana_res['ct_pair'] = liana_res[source_key] + cell_separator + liana_res[target_key]
+    liana_res['interaction'] = liana_res[ligand_key] + lr_sep + liana_res[receptor_key]
+    liana_res['ct_pair'] = liana_res[source_key] + cell_sep + liana_res[target_key]
     liana_res = liana_res[[sample_key, 'ct_pair', 'interaction', score_key]]
     
     # get scores & invert if necessary
@@ -266,7 +256,7 @@ def lrs_to_views(adata,
                                     columns=sample_key,
                                     values=score_key)
     
-        lrs_wide.index = view + var_separator + lrs_wide.index
+        lrs_wide.index = view + var_sep + lrs_wide.index
         lrs_wide = lrs_wide.replace(np.nan, lr_fill)
         
         if lrs_wide.shape[0] >= lrs_per_view: # check if enough LRs
@@ -279,7 +269,7 @@ def lrs_to_views(adata,
                 lr_adatas[view] = temp
                 
     # to mdata
-    mdata = mu.MuData(lr_adatas)
+    mdata = MuData(lr_adatas)
     
     # process metadata
     _process_meta(adata=adata, mdata=mdata, sample_key=sample_key, obs_keys=obs_keys)
@@ -295,7 +285,7 @@ def _dataframe_to_anndata(df):
     return AnnData(X=X, obs=obs, var=var, dtype=np.float32)
 
 
-def _remove_mod_var(mdata, markers, view_separator, var_column):
+def _remove_mod_var(mdata, markers, view_sep, var_column):
     for current_mod in mdata.mod.keys():
         # markers in markers dict for each modality except for current_mod
         negative_markers = [marker for mod in markers.keys() if mod != current_mod for marker in markers[mod]]
@@ -303,8 +293,8 @@ def _remove_mod_var(mdata, markers, view_separator, var_column):
         if current_mod not in list(markers.keys()):
             warnings.warn('no markers in dict for view: {0}'.format(current_mod), Warning)
         else:
-            #keep negative_markers not in markers[current_mod] and add view_separator
-            negative_markers = [current_mod + view_separator + marker for marker in negative_markers if marker not in markers[current_mod]]
+            #keep negative_markers not in markers[current_mod] and add view_sep
+            negative_markers = [current_mod + view_sep + marker for marker in negative_markers if marker not in markers[current_mod]]
         
         if var_column is None:
             # remove negative_markers from current_mod
@@ -315,11 +305,12 @@ def _remove_mod_var(mdata, markers, view_separator, var_column):
 
     mdata.update()
 
-def filter_view_markers(mdata,
-                        markers,
-                        view_separator=':',
-                        var_column='highly_variable',
-                        inplace=False
+@d.dedent
+def filter_view_markers(mdata: MuData,
+                        markers: dict,
+                        view_sep: str = ':',
+                        var_column: str = 'highly_variable',
+                        inplace: bool =False
                         ):
     """
     Used for removing potential cell type marker genes found in the background of other views and thought to be contamination.
@@ -328,17 +319,15 @@ def filter_view_markers(mdata,
 
     Parameters
     ----------
-    mdata :class:`~mudata.MuData`
-        MuData object. Highly variable genes should be computed already in .var for each modality.
+    %(mdata)s
     markers :class:`dict`
         Dictionary with markers for each view. Keys are the views and values are lists of markers. Can contain markers for views that are not in mdata.mod.keys().
-    view_separator :class:`str`, optional
+    view_sep :class:`str`, optional
         Separator between view and gene names. Defaults to ':'.
     var_column :class:`str`, optional
         Column in mdata.mod['some_view'].var that contains the highly variable genes. Defaults to 'highly_variable'.
         If set to ``None``, instead of setting the hvg genes to False, the hvg genes will be removed from the view.
-    inplace :class:`bool`, optional
-        If True, update mdata in place, else makes a copy. Defaults to False.
+    %(inplace)s
     """
     # check if markers is a dict
     if not isinstance(markers, dict):
@@ -354,12 +343,11 @@ def filter_view_markers(mdata,
             raise ValueError('{0} is not in the columns of .var for all modalities'.format(var_column))
 
     if inplace:
-        _remove_mod_var(mdata, markers, view_separator, var_column)
+        _remove_mod_var(mdata, markers, view_sep, var_column)
     else:
         cdata = mdata.copy()
-        _remove_mod_var(cdata, markers, view_separator, var_column)
+        _remove_mod_var(cdata, markers, view_sep, var_column)
         return cdata
-
 
 
 def _process_meta(adata, mdata, sample_key, obs_keys):