+++ liana/utils/mdata_to_anndata.py
@@ -1,6 +1,8 @@
 import anndata as an
 from liana.method._pipe_utils._pre import _choose_mtx_rep
+from liana._constants._docs import d
 
+@d.dedent
 def mdata_to_anndata(mdata,
                      x_mod, y_mod,
                      x_layer=None, y_layer=None,
@@ -15,7 +17,7 @@ def mdata_to_anndata(mdata,
     Parameters
     ----------
     mdata
-        MuData object to be converted.
+        MuData object.
     x_mod
         Name of the modality to be used as x.
     y_mod
@@ -32,8 +34,7 @@ def mdata_to_anndata(mdata,
         Transformation function to be applied to modality x.
     y_transform
         Transformation function to be applied to modality y.
-    verbose
-        Whether to print progress.
+    %(verbose)s
     
     Returns
     -------
@@ -41,14 +42,11 @@ def mdata_to_anndata(mdata,
     Information related to observations (obs, obsp, obsm) and `.uns` are copied from the original MuData object.
     """
 
-    xdata, ydata = _handle_mdata(mdata, 
-                                 x_mod, y_mod,
-                                 x_layer, y_layer,
-                                 x_use_raw, y_use_raw,
-                                 x_transform=x_transform,
-                                 y_transform=y_transform,
-                                 verbose=verbose,
-                                 )
+    if x_mod is None or y_mod is None:
+        raise ValueError("Both `x_mod` and `y_mod` must be provided!")
+    
+    xdata = _handle_mod(mdata, x_mod, x_use_raw, x_layer, x_transform, verbose)
+    ydata = _handle_mod(mdata, y_mod, y_use_raw, y_layer, y_transform, verbose)
     
     adata = an.concat([xdata, ydata], axis=1, label='modality')
     
@@ -58,23 +56,6 @@ def mdata_to_anndata(mdata,
     adata.uns = mdata.uns.copy()
     
     return adata
-
-
-# TODO: Do I still use this anywhere?
-# If not, merge with mdata_to_anndata & _handle_mod
-def _handle_mdata(mdata, 
-                  x_mod, y_mod,
-                  x_layer, y_layer,
-                  x_use_raw, y_use_raw,
-                  x_transform, y_transform, 
-                  verbose):
-    if x_mod is None or y_mod is None:
-        raise ValueError("Both `x_mod` and `y_mod` must be provided!")
-    
-    xdata = _handle_mod(mdata, x_mod, x_use_raw, x_layer, x_transform, verbose)
-    ydata = _handle_mod(mdata, y_mod, y_use_raw, y_layer, y_transform, verbose)
-    
-    return xdata, ydata
 
 def _handle_mod(mdata, mod, use_raw, layer, transform, verbose):
     if mod not in mdata.mod.keys():